requirements.txt
fastapi[all]~=0.115.0
pymongo~=4.10.1
pydantic~=2.9.2
requests~=2.32.3
python-dotenv~=1.0.1